Add fn to get existing packaged entitlements
If the entitlements already exist somewhere within a Payload, it may
cause signing issues - despite a valid entitlements.plist passed during
codesign.

Specifically, this is an issue with Swift apps (or at least one app in
particular) which isn't valid unless the internal entitlements are
updated.

There's logic here to search the Payload .app dir, and one level up
(where the Payload was extracted to). A search is performed one level up
because there was an IPA with an entitlements packaged at the root.
It's not clear if and when this is necessary.