Add support for generating entitlements.xcent file
In some cases, there's a file called entitlements.xcent, which contains
similar (or identical?) entitlements as the actual entitlements reported
by codesign. But it's not in sync with codesign - which will report the
entitlements the app was signed with, not the contents of this xcent
file.

So it seems like this file is actually just some artifact, and changing
it has no impact on what the signature of the app actually is. At least
from the perspective of codesign - which should be the single source of
truth (right?).

Regardless, just in case, it seems there's no harm just copying this
file over to the resigned version if it exists.
Add support for generating entitlements.xcent file