Change sign_payload to use hardcoded Frameworks dir
Fixes an issue where payload app directories with spaces wouldn't be
signed because of unescaped spaces. This way the directory is always
quoted, and the Frameworks are globbed correctly.

Besides, it makes sense for sign_payload to be opinionated about the
Frameworks directory, which regardless is consistent with the rest of
the script and its knowledge of the domain.